nouveau: defer engine sync to CPU access and stream uploads through a GART ring
Wanderer streams background images one 1920x1 row per PutImage, and
every row paid a GART wait, a mapping wait, and two fence round trips -
around 2ms per row, seconds per screen. Accelerated fills, blits and
uploads now mark the bitmap dirty instead of waiting, and the CPU
settles the pending work only when it actually touches a mapping
(MAP_BUFFER, the PutPixel/GetPixel fast paths, and before an engine
download). The GART staging buffer becomes a ring so consecutive
uploads stream into it, waiting only when the ring wraps. The same
measured workload dropped from 2364ms to 24ms.

This also closes a pre-existing hazard on every generation: the
FillRect path issued engine work and returned without ever waiting, so
a queued fill could destroy CPU-rendered content that followed it.

The GART size line printed through a u32 - a 2^57-byte VMM limit came
out as "0 MiB". Print the real value.

+++ workbench/hidds/nouveau/drm/nouveau/nouveau_ttm.c
@@ -341,7 +341,7 @@ nouveau_ttm_init(struct nouveau_drm *drm)
 	INIT_LIST_HEAD(&drm->ttm.io_reserve_lru);
 
 	NV_INFO(drm, "VRAM: %d MiB\n", (u32)(drm->gem.vram_available >> 20));
-	NV_INFO(drm, "GART: %d MiB\n", (u32)(drm->gem.gart_available >> 20));
+	NV_INFO(drm, "GART: %llu MiB\n", (unsigned long long)(drm->gem.gart_available >> 20));
 	return 0;
 }
